feat: adds audio format as an envar
