+++ .env.example
@@ -2,6 +2,7 @@
 ASSISTANT_TOOL=rasa # Define the assistant tool to be used. Options: rasa, anthropic
 STT_TOOL=whisper-local # Define the STT tool to be used. Options: whisper-local, whisper
 SQL_DB_FILE_NAME="freetalkbot.db" # Name of the SQLite database file to be used by the whatsapp bot
+AUDIO_FORMAT=pcm16 # Audio format that will use audiosocket server. Options: pcm16, g711
 
 # Rasa variables. Mandatory if ASSISTANT_TOOL=rasa.
 # Used in rasa implementation and in golang communication channels
@@ -23,5 +24,6 @@ WHISPER_LOCAL_URL=whisper_cpu:8000/v1 # Mandatory if STT_TOOL=whisper-local
 WHISPER__MODEL="deepdml/faster-whisper-large-v3-turbo-ct2" # The whisper model to use. Mandatory if STT_TOOL=whisper-local.
 
 # Optional variables
+G711_AUDIO_CODEC=ulaw # Audio codec to be used in g711 audio format. Options: ulaw, alaw
 #PAIR_PHONE_NUMBER=+1234567890 # Use this variable to allow pair your whatsapp account with a pairing code
 #LOG_LEVEL=DEBUG  # Use this variable to enable debug logs
